AX: ActiveElementChanged notifications for combobox should target the combobox in question, not its controlled or owned element. https://bugs.webkit.org/show_bug.cgi?id=274178 <rdar://problem/128083384>
Reviewed by NOBODY (OOPS!).

Until now, if a combobox controlled or owned another element, we imposed the rule that the target of the ActiveElementChanged notification should be the controlled or owned element, typically a list, listbox, tree or grid. This rule is unnecessary and adds significant complexity to the handling of these changes. Furthermore, this is the root cause of problems in comboboxes like the search field in google.com. This patch eliminates this rule and simplifies the code around combobox ActiveElementChanged notifications.

* LayoutTests/accessibility/combobox/aria-combobox-control-owns-elements-expected.txt:
* LayoutTests/accessibility/combobox/aria-combobox-control-owns-elements.html:
* LayoutTests/accessibility/combobox/aria-combobox-no-owns-expected.txt:
* LayoutTests/accessibility/combobox/aria-combobox-no-owns.html:
* LayoutTests/accessibility/combobox/mac/combobox-activedescendant-notifications.html:
* Source/WebCore/accessibility/AXCoreObject.cpp:
(WebCore::AXCoreObject::activeDescendant const):
(WebCore::AXCoreObject::canBeControlledBy const): Deleted.
* Source/WebCore/accessibility/AXCoreObject.h:
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::handleActiveDescendantChange):
* Source/WebCore/accessibility/atspi/AXObjectCacheAtspi.cpp:
(WebCore::AXObjectCache::postPlatformNotification):
* Source/WebCore/accessibility/mac/AXObjectCacheMac.mm:
(WebCore::AXObjectCache::postPlatformNotification):